requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
discord-py-interactions==4.3.4
frozenlist==1.3.1
idna==3.4
interactions-lavalink==0.1.2
lavalink==4.0.4
multidict==6.0.2
yarl==1.8.1
